Fix printf format string specifier for R_xlen_t in error message for a signed type.
git-svn-id: https://svn.r-project.org/R/trunk@74620 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/memory.c
@@ -4004,7 +4004,7 @@ void(SET_STRING_ELT)(SEXP x, R_xlen_t i, SEXP v)
     if (TYPEOF(CHK(v)) != CHARSXP)
         error("Value of SET_STRING_ELT() must be a 'CHARSXP' not a '%s'", type2char(TYPEOF(v)));
     if (i < 0 || i >= XLENGTH(x))
-        error(_("attempt to set index %lu/%lu in SET_STRING_ELT"), i, XLENGTH(x));
+        error(_("attempt to set index %lld/%lld in SET_STRING_ELT"), (long long)i, (long long)XLENGTH(x));
     CHECK_OLD_TO_NEW(x, v);
     if (ALTREP(x))
         ALTSTRING_SET_ELT(x, i, v);
@@ -4024,7 +4024,7 @@ SEXP(SET_VECTOR_ELT)(SEXP x, R_xlen_t i, SEXP v)
         error("%s() can only be applied to a '%s', not a '%s'", "SET_VECTOR_ELT", "list", type2char(TYPEOF(x)));
     }
     if (i < 0 || i >= XLENGTH(x))
-        error(_("attempt to set index %lu/%lu in SET_VECTOR_ELT"), i, XLENGTH(x));
+        error(_("attempt to set index %lld/%lld in SET_VECTOR_ELT"), (long long)i, (long long)XLENGTH(x));
     FIX_REFCNT(x, VECTOR_ELT(x, i), v);
     CHECK_OLD_TO_NEW(x, v);
     return VECTOR_ELT(x, i) = v;
